Modals: Use Bootstrap modals instead of custom solution
They just work better.
