+++ app/styles/main.scss
@@ -12,6 +12,7 @@
 @import 'utilities';
 @import './../../node_modules/bootstrap/scss/utilities/api';
 
+@import './../../node_modules/bootstrap/scss/modal';
 
 // BASE
 @import 'base/globals';
@@ -41,7 +42,7 @@
 @import 'modules/cards';
 @import 'modules/lists';
 @import 'modules/tabs';
-@import 'modules/modals';
+//@import 'modules/modals';
 @import 'modules/ratings';
 @import 'modules/review';
 @import 'modules/badges';
