+++ app/styles/modules/_tabs.scss
@@ -90,11 +90,11 @@
       &:after {
         content: '';
         position: absolute;
-        bottom: -(_map-deep-get($tabs-card, 'default', margin-bottom) + (_get-rem($default-border-radius)));
+        bottom: -(_map-deep-get($tabs-card, 'default', margin-bottom) + $default-border-radius);
         left: -1px;
         right: -1px;
 
-        height: _map-deep-get($tabs-card, 'default', margin-bottom) + _get-rem($default-border-radius * 2) + _get-rem(3px);
+        height: _map-deep-get($tabs-card, 'default', margin-bottom) + $default-border-radius * 2 + _get-rem(3px);
         background-color: _map-deep-get($tabs-card, 'active', background-color);
         border: _map-deep-get($tabs-card, 'default', border);
         border-color: _map-deep-get($tabs-card, 'active', border-color);
