+++ app/styles/utils/_variables.scss
@@ -61,7 +61,10 @@ $shadow-lg: $shadow-direct-5;
 // ----------------------------
 // VARIOUS
 
-$border-radius: 3px !default;
+$border-radius: 0.25rem !default;
+$border-radius-sm: 0.2rem !default;
+$border-radius-lg: 0.3rem !default;
+$border-radius-pill: 50rem !default;
 $border-radius-circle: 999rem !default;
 
 // ---------------------------------------------
@@ -73,6 +76,9 @@ $z-index-topbar: 1400 !default;
 $z-index-dropdown: 1450 !default;
 $z-index-overlay: 1500 !default;
 $z-index-alert: 1999 !default;
+
+$zindex-modal-backdrop: 1500 !default;
+$zindex-modal: 1550 !default;
 
 // ---------------------------------------------
 // EASINGS
@@ -89,6 +95,10 @@ $ease-100-delayed: 0.1s ease-out 0.1s !default;
 $ease-600: 0.6s ease 0s !default;
 
 $slow-end-anim: cubic-bezier(0.25, 0.25, 0, 0.995) !default;
+
+$enable-transitions: true !default;
+$enable-reduced-motion: true !default;
+$enable-shadows: true !default;
 
 // BOOTSTRAP SETTINGS
