+++ app/views/styleguide/data/sitemap.yaml
@@ -267,6 +267,9 @@ l_sitemap:
             title: "Modals"
             href: "modals"
           -
+            title: "Legacy Styleguide Modals"
+            href: "legacy-modals"
+          -
             title: "Multiple modals"
             href: "multiple-modals"
       -
